requirements.txt
django==5.1.1
psycopg2==2.9.9
python-dotenv